fix: Use compact plans during merge
