--- internal_types/src/arrow.rs
@@ -1,1 +1,0 @@
-pub mod sort;
