+++ internal_types/src/lib.rs
@@ -6,7 +6,6 @@
     clippy::clone_on_ref_ptr
 )]
 
-pub mod arrow;
 pub mod once;
 pub mod schema;
 pub mod selection;
