+++ query_tests/src/sql.rs
@@ -319,9 +319,9 @@ async fn sql_select_from_system_chunk_columns() {
         "+---------------+----------+------------+-------------+-------------------+-------+-----------+-----------+-----------------+",
         "| partition_key | chunk_id | table_name | column_name | storage           | count | min_value | max_value | estimated_bytes |",
         "+---------------+----------+------------+-------------+-------------------+-------+-----------+-----------+-----------------+",
-        "| 1970-01-01T00 | 0        | h2o        | city        | ReadBuffer        | 2     | Boston    | Boston    | 247             |",
+        "| 1970-01-01T00 | 0        | h2o        | city        | ReadBuffer        | 2     | Boston    | Boston    | 252             |",
         "| 1970-01-01T00 | 0        | h2o        | other_temp  | ReadBuffer        | 1     | 70.4      | 70.4      | 369             |",
-        "| 1970-01-01T00 | 0        | h2o        | state       | ReadBuffer        | 2     | MA        | MA        | 235             |",
+        "| 1970-01-01T00 | 0        | h2o        | state       | ReadBuffer        | 2     | MA        | MA        | 240             |",
         "| 1970-01-01T00 | 0        | h2o        | temp        | ReadBuffer        | 1     | 70.4      | 70.4      | 369             |",
         "| 1970-01-01T00 | 0        | h2o        | time        | ReadBuffer        | 2     | 50        | 250       | 51              |",
         "| 1970-01-01T00 | 0        | o2         | city        | OpenMutableBuffer | 1     | Boston    | Boston    | 35              |",
