+++ server/src/db/lifecycle.rs
@@ -7,6 +7,10 @@ use data_types::chunk_metadata::{ChunkAddr, ChunkStorage};
 use data_types::database_rules::{LifecycleRules, SortOrder};
 use data_types::error::ErrorLogger;
 use data_types::job::Job;
+use data_types::partition_metadata::{InfluxDbType, TableSummary};
+use hashbrown::HashMap;
+use internal_types::schema::sort::SortKey;
+use internal_types::schema::TIME_COLUMN_NAME;
 use lifecycle::{
     ChunkLifecycleAction, LifecycleChunk, LifecyclePartition, LifecycleReadGuard,
     LifecycleWriteGuard, LockableChunk, LockablePartition,
@@ -225,4 +229,33 @@ impl LifecycleChunk for CatalogChunk {
     fn row_count(&self) -> usize {
         self.storage().0
     }
+}
+
+/// Compute a sort key that orders lower cardinality columns first
+///
+/// In the absence of more precise information, this should yield a
+/// good ordering for RLE compression
+fn compute_sort_key<'a>(summaries: impl Iterator<Item = &'a TableSummary>) -> SortKey<'a> {
+    let mut cardinalities: HashMap<&str, u64> = Default::default();
+    for summary in summaries {
+        for column in &summary.columns {
+            if column.influxdb_type != Some(InfluxDbType::Tag) {
+                continue;
+            }
+
+            if let Some(count) = column.stats.distinct_count() {
+                *cardinalities.entry(column.name.as_str()).or_default() += count.get()
+            }
+        }
+    }
+
+    let mut cardinalities: Vec<_> = cardinalities.into_iter().collect();
+    cardinalities.sort_by_key(|x| x.1);
+
+    let mut key = SortKey::with_capacity(cardinalities.len() + 1);
+    for (col, _) in cardinalities {
+        key.push(col, Default::default())
+    }
+    key.push(TIME_COLUMN_NAME, Default::default());
+    key
 }
