+++ server/src/db/lifecycle/compact.rs
@@ -4,12 +4,8 @@ use std::future::Future;
 use std::sync::Arc;
 
 use futures::StreamExt;
-use hashbrown::HashMap;
 
 use data_types::job::Job;
-use data_types::partition_metadata::{InfluxDbType, TableSummary};
-use internal_types::schema::sort::SortKey;
-use internal_types::schema::TIME_COLUMN_NAME;
 use lifecycle::LifecycleWriteGuard;
 use query::frontend::reorg::ReorgPlanner;
 use query::QueryChunkMeta;
@@ -20,36 +16,8 @@ use crate::db::catalog::chunk::CatalogChunk;
 use crate::db::catalog::partition::Partition;
 use crate::db::DbChunk;
 
+use super::compute_sort_key;
 use super::{error::Result, LockableCatalogChunk, LockableCatalogPartition};
-
-/// Compute a sort key that orders lower cardinality columns first
-///
-/// In the absence of more precise information, this should yield a
-/// good ordering for RLE compression
-fn compute_sort_key<'a>(summaries: impl Iterator<Item = &'a TableSummary>) -> SortKey<'a> {
-    let mut cardinalities: HashMap<&str, u64> = Default::default();
-    for summary in summaries {
-        for column in &summary.columns {
-            if column.influxdb_type != Some(InfluxDbType::Tag) {
-                continue;
-            }
-
-            if let Some(count) = column.stats.distinct_count() {
-                *cardinalities.entry(column.name.as_str()).or_default() += count.get()
-            }
-        }
-    }
-
-    let mut cardinalities: Vec<_> = cardinalities.into_iter().collect();
-    cardinalities.sort_by_key(|x| x.1);
-
-    let mut key = SortKey::with_capacity(cardinalities.len() + 1);
-    for (col, _) in cardinalities {
-        key.push(col, Default::default())
-    }
-    key.push(TIME_COLUMN_NAME, Default::default());
-    key
-}
 
 /// Compact the provided chunks into a single chunk,
 /// returning the newly created chunk
