+++ server/src/db/lifecycle/move_chunk.rs
@@ -1,10 +1,11 @@
-use crate::db::catalog::chunk::CatalogChunk;
 pub(crate) use crate::db::chunk::DbChunk;
+use crate::db::{catalog::chunk::CatalogChunk, lifecycle::compute_sort_key};
 use ::lifecycle::LifecycleWriteGuard;
 use data_types::job::Job;
-use internal_types::{arrow::sort::sort_record_batch, selection::Selection};
+use futures::StreamExt;
 
 use observability_deps::tracing::{debug, info};
+use query::{frontend::reorg::ReorgPlanner, QueryChunkMeta};
 use read_buffer::{ChunkMetrics as ReadBufferChunkMetrics, RBChunk};
 use std::{future::Future, sync::Arc};
 use tracker::{TaskTracker, TrackedFuture, TrackedFutureExt};
@@ -33,10 +34,11 @@ pub fn move_chunk_to_read_buffer(
 
     // update the catalog to say we are processing this chunk and
     // then drop the lock while we do the work
-    let (mb_chunk, table_summary) = {
-        let mb_chunk = guard.set_moving(&registration)?;
-        (mb_chunk, guard.table_summary())
-    };
+    guard.set_moving(&registration)?;
+    let table_summary = guard.table_summary();
+
+    // snapshot the data
+    let query_chunks = vec![DbChunk::snapshot(&*guard)];
 
     // Drop locks
     let chunk = guard.unwrap().chunk;
@@ -51,20 +53,24 @@ pub fn move_chunk_to_read_buffer(
         ReadBufferChunkMetrics::new(&metrics, db.catalog.metrics().memory().read_buffer()),
     );
 
+    let ctx = db.exec.new_context();
+
     let fut = async move {
         info!(chunk=%addr, "chunk marked MOVING, loading tables into read buffer");
 
-        // load table into the new chunk one by one.
-        debug!(chunk=%addr, "loading table to read buffer");
-        let batch = mb_chunk
-            .read_filter(Selection::All)
-            // It is probably reasonable to recover from this error
-            // (reset the chunk state to Open) but until that is
-            // implemented (and tested) just panic
-            .expect("Loading chunk to mutable buffer");
+        let key = compute_sort_key(query_chunks.iter().map(|x| x.summary()));
 
-        let sorted = sort_record_batch(batch).expect("failed to sort");
-        rb_chunk.upsert_table(&table_summary.name, sorted);
+        // Cannot move query_chunks as the sort key borrows the column names
+        let (_schema, plan) =
+            ReorgPlanner::new().compact_plan(query_chunks.iter().map(Arc::clone), key)?;
+
+        let physical_plan = ctx.prepare_plan(&plan)?;
+        let mut stream = ctx.execute(physical_plan).await?;
+
+        // Collect results into RUB chunk
+        while let Some(batch) = stream.next().await {
+            rb_chunk.upsert_table(&table_summary.name, batch?)
+        }
 
         // Can drop and re-acquire as lifecycle action prevents concurrent modification
         let mut guard = chunk.write();
